requirements.txt
Django~=3.1.7
djangorestframework~=3.12.2
djangorestframework-simplejwt~=4.6.0
django-cors-headers~=3.7.0
django-perms~=2.1.0
mysqlclient~=2.0.3